Refactor fvc matrices to fvc.js
