+++ js/chart/chart.js
@@ -32,8 +32,9 @@ import {div, el, fragment, label, textNode} from './react-hyper';
 import classNames from 'classnames';
 import {isSet, bitCount} from '../models/bitmap';
 import {xenaColor} from '../xenaColor';
+var fvc = require('./fvc');
 var {groupValues} = require('./dataUtils');
-var {applyExpression, computeAvgExpr, computePctExpr} = require('./singleCell');
+var {applyExpression} = require('./singleCell');
 var {fastats} = require('../xenaWasm');
 var {reOrderFields} = require('../models/denseMatrix');
 
@@ -430,96 +431,6 @@ const LOWER = 1;
 const MEDIAN = 2;
 const UPPER = 3;
 const UPPERWHISKER = 4;
-const BOXLEN = 5;
-
-function boxplotPoint(data) {
-	var m = data.length;
-	data.sort((a, b) => a - b);
-
-	// http://onlinestatbook.com/2/graphing_distributions/boxplots.html
-	var median = data[Math.floor( m / 2)],
-		lower =  data[Math.floor( m / 4)],
-		upper =  data[Math.floor( 3 * m / 4)],
-		whisker = 1.5 * (upper - lower),
-		upperwhisker = data[
-			_.findIndexDefault(data, x => x > upper + whisker, data.length) - 1],
-		lowerwhisker = data[
-			_.findLastIndexDefault(data, x => x < lower - whisker, -1) + 1];
-
-	// This must match BOX order
-	return [lowerwhisker, lower, median, upper, upperwhisker];
-}
-
-// poor man's lazy seq
-function* constantly (fn) {
-	while (true) {
-		yield fn();
-	}
-}
-
-var nobox = new Array(BOXLEN).fill(NaN);
-
-function initFVCMatrices({ydata, groups, yexpression, ynonexpressed}) {
-	// matrices, row is x and column is y:
-	// mean
-	// median
-	// upper --  75 percentile
-	// lower --  25 percentile
-	// upperwhisker --  upperwhisker percentile
-	// lowerwhisker --  lowerwhisker percentile
-	// std
-	// nNumber -- number of data points (real data points) for dataMatrix
-
-	// init average matrix std matrix // row is x by column y
-	var [meanMatrix, stdMatrix, nNumberMatrix, expressionMatrix, detectionMatrix] =
-			constantly(() =>
-				_.times(groups.length, () => new Array(ydata.length).fill(NaN))),
-		boxes = _.times(groups.length, () => new Array(ydata.length));
-
-	var isSingleCell = yexpression === 'singleCell';
-
-	// Y data and fill in the matrix
-	ydata.forEach((ydataElement, k) => {
-		// Single cell mode (dot plot only):
-		// Retrieve non-expressed indices for the ydata element, and filter those indices from groups, then compute the binned values.
-		// Track how many non-expressed samples were removed from each group.
-		// Bulk mode:
-		// Just compute the binned values.
-		let nonExpressedSet = ynonexpressed.get(k),
-			nonExpressedCountByGroup = new Map(),
-			expressedGroupsOrGroups = isSingleCell ? _.map(groups, (group, g) => {
-				nonExpressedCountByGroup.set(g, 0);
-				return _.filter(group, i => {
-					if (nonExpressedSet.has(i)) {
-						nonExpressedCountByGroup.set(g, nonExpressedCountByGroup.get(g) + 1);
-						return false;
-					}
-					return true;
-				});
-			}) : groups,
-			ybinnedSample = groupValues(ydataElement, expressedGroupsOrGroups);
-
-		// Note that xCategories has already been null filtered on x, so it's not
-		// the same as xcodemap.
-		ybinnedSample.forEach((data, i) => {
-			let m = data.length;
-			nNumberMatrix[i][k] = m;
-			boxes[i][k] = m ? boxplotPoint(data) : nobox;
-			if (m) {
-				let average =  highchartsHelper.average(data);
-				meanMatrix[i][k] = average;
-				stdMatrix[i][k] = highchartsHelper.standardDeviation(data, average);
-				if (isSingleCell) {
-					let nonExpressedCount = nonExpressedCountByGroup.get(i) || 0,
-						totalCount = m + nonExpressedCount;
-					detectionMatrix[i][k] = computePctExpr(m, totalCount);
-					expressionMatrix[i][k] = computeAvgExpr(data);
-				}
-			}
-		});
-	});
-	return {detectionMatrix, expressionMatrix, meanMatrix, boxes, stdMatrix, nNumberMatrix};
-}
 
 function sortMatrices(xCategories, groups, colors, matrices) {
 	let {boxes} = matrices,
@@ -744,7 +655,7 @@ var fvcChart = chartType => ({
 function boxOrDotOrViolin({groups, xCategories, chartType = 'boxplot', colors, inverted, setHasStats, setView, yexpression, yfields, ydata,
 		xlabel, ylabel, ynonexpressed, ynorm}, chartOptions) {
 	setView(chartType);
-	var matrices = initFVCMatrices({ydata, groups, yexpression, ynonexpressed});
+	var matrices = fvc.getMatrices({ydata, groups, yexpression, ynonexpressed});
 
 	// sort by median of xCategories if yfields.length === 1
 	if (xCategories.length > 0 && yfields.length === 1) {
@@ -1439,7 +1350,7 @@ class Chart extends PureComponent {
 			xexp = xexpOpts[chartState.expState[xcolumn]],
 			yexp = yexpOpts[chartState.expState[ycolumn]],
 			ymin = yMin(ydata0),
-			// 'singleCell' expression mode only for dot plots with positive ydata
+			// 'singleCell' expression mode only for dot plots with positive values
 			yexpression = expressionMode(chartState, ymin),
 			ynorm = !ycodemap && _.get(normalizationOptions[
 					chartState.normalizationState[chartState.ycolumn]], 'value'),
